+++ app/helpers/pages_core/plugin_assets_helper.rb
@@ -1,8 +1,0 @@
-# encoding: utf-8
-
-module PagesCore::PluginAssetsHelper
-
-  include ActionView::Helpers::AssetTagHelper
-  #include Sprockets::Helpers
-
-end